feat(data.elm update.elm): automatic motion of enemy, change turn, define distance and related func
Realize the change of the turn between the player and the automatic enemy. Realize the automatic
motion of the enemy. (Not that intelligent now. Try to move towards the nearest hero 2 step each
turn without bypassing barriers) Define the distance between two Hexagons and define function to
calculate distance between two Hexagons and function to find the least distance between one Hexagon
and a list of Hexagons.

re #5