feat(msg, model, update, updateshop, view, viewshop): add shop UI, upgrade and draw system
In the shop, press ENTER to enter the buyItem page, in which the player can draw a new hero or
upgrade heroes. In upgrade hero page, press ← or → to switch heroes. If you don't have one hero, the
picture will be a locked hero. The gradient light effect and hero information is to be added

Beta Feature: Shop UI System #183